Adds parser for SMS commands
The parser is designed to handle four rather simple commands:
* create
* edit
* checkin
* complete

The first three commands contain Points, a geo-located data-point.
These can contain more data than just the coordinates, but
doesn't have to. These data are:
* Action
  * Tent
  * Hut
  * Food
* Message
* Time
  * DateTime
  * Date
  * Time

The parsers for coordinates and timestamps are very forgiving,
so there are a range of ways to enter the same data, e.g.
* 7428600
* 7428600.365
* N74286
* N74286.365

Coordinates must have at least three digits to be auto-completed.
This is to avoid confusion with the UTM-zones that can be
expressed as:
* UTM33
* 33

An example of a complete and valid input is:
create
33 E618 7428602.235 2020-11-10 07:12:36 tent
"Start of hike! Getting of the bus here."